requirements.txt
streamlit
pandas
numpy
openpyxl
pillow